Added route component
