+++ src/App.js
@@ -1,38 +1,13 @@
 import React from 'react';
-import CategoryDirectory from './components/category-directory/category-directory.component';
+import Home from './routes/home/home.component';
+import { Routes, Route } from 'react-router-dom';
 
 const App = () => {
 
-  const categories = [
-    {
-      id: 1,
-      title: 'Hats',
-      imageUrl: 'https://i.ibb.co/cvpntL1/hats.png',
-    },
-    {
-      id: 2,
-      title: 'Jackets',
-      imageUrl: 'https://i.ibb.co/px2tCc3/jackets.png',
-    },
-    {
-      id: 3,
-      title: 'Sneakers',
-      imageUrl: 'https://i.ibb.co/0jqHpnp/sneakers.png',
-    },
-    {
-      id: 4,
-      title: 'Women',
-      imageUrl: 'https://i.ibb.co/GCCdy8t/womens.png',
-    },
-    {
-      id: 5,
-      title: 'Men',
-      imageUrl: "https://i.ibb.co/R70vBrQ/men.png",
-    },
-  ]
-
   return (
-    <CategoryDirectory categories={categories}/>
+    <Routes>
+      <Route path='/' element={<Home />} />
+    </Routes>
   );
 }
